Extract to shared helper
